Invoke resolvers with custom messages and introduce SyncResolver
